chore: move integration tests to separate file and create a separate test for SQLite with foreign key constraint enabled
